Add: Parser static class. I plan to move all parsing to this class instead of containing them inside models
