+++ Poushec.UpdateCatalogParser/Models/CatalogSearchResult.cs
@@ -14,7 +14,7 @@ namespace Poushec.UpdateCatalogParser.Models
         public int SizeInBytes { get; set; }
         public string UpdateID { get; set; }
 
-        private CatalogSearchResult(
+        internal CatalogSearchResult(
             string title, 
             string products, 
             string classification, 
@@ -34,20 +34,5 @@ namespace Poushec.UpdateCatalogParser.Models
             this.UpdateID = updateId;
         }
 
-        public static CatalogSearchResult ParseFromResultsTableRow(HtmlNode resultsRow)
-        {
-            HtmlNodeCollection rowCells = resultsRow.SelectNodes("td");
-            
-            string title = rowCells[1].InnerText.Trim();
-            string products = rowCells[2].InnerText.Trim();
-            string classification = rowCells[3].InnerText.Trim();
-            DateTime lastUpdated = DateTime.Parse(rowCells[4].InnerText.Trim());
-            string version = rowCells[5].InnerText.Trim();
-            string size = rowCells[6].SelectNodes("span")[0].InnerText;
-            int sizeInBytes = int.Parse(rowCells[6].SelectNodes("span")[1].InnerHtml);
-            string updateID = rowCells[7].SelectNodes("input")[0].Id;
-
-            return new CatalogSearchResult(title, products, classification, lastUpdated, version, size, sizeInBytes, updateID);
-        }
     }
 }
